refactor: move `CreateNewListModal`
